Majority of prototype
This is the initial push for this project but it's pretty much the majority of the prototype version completed. Only 2-3 activities left with GUIs to be completed. Will be adding some graphs too for some nice visuals. The symbol PNGs and sample data lists will be getting removed. There's essentially no function yet so I'll be adding in Retrofit/OkHttp once the web service prototype is ready