Linking to web service
I more or less finished the web service last night so this commit is linking the app to that api. Enabled user login, registration, and retrieval of sensor data. Added dependencies for Retrofit, OkHttp and Gson to handle the networking. Still some work to do but it's in a much better place now for the prototype.

Fixed a load of warnings involving deprecated methods, unused hardcoded data, and unused imports.